requirements.txt
flask
pdfplumber
python-docx
playwright
google-generativeai